requirements.txt
streamlit
google-cloud-vision
pandas
openpyxl